created a new file for utils funcs
