+++ pennylane/pauli/__init__.py
@@ -17,7 +17,6 @@ from .pauli_arithmetic import PauliWord, PauliSentence
 
 from .utils import (
     is_pauli_word,
-    pauli_word_prefactor,
     are_identical_pauli_words,
     pauli_to_binary,
     binary_to_pauli,
@@ -38,6 +37,8 @@ from .utils import (
     diagonalize_qwc_groupings,
     simplify,
 )
+
+from .pauli_interface import pauli_word_prefactor
 
 from .conversion import (
     pauli_decompose,
